refactor: extraire BasePageChantierLayout
